% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ir.R
\name{build_ir}
\alias{build_ir}
\title{Build the lineage IR from parsed SQL}
\usage{
build_ir(parsed)
}
\arguments{
\item{parsed}{The list returned by \code{\link[=parse_sql]{parse_sql()}}.}
}
\value{
An object of class \code{rdataflow_ir} (a list of tibbles; see file
header for the schema).
}
\description{
Build the lineage IR from parsed SQL
}
