% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_transform.R
\name{classify_transform}
\alias{classify_transform}
\title{Classify the transformations in a lineage IR}
\usage{
classify_transform(ir)
}
\arguments{
\item{ir}{An object of class \code{rdataflow_ir} (see \code{\link[=build_ir]{build_ir()}}).}
}
\value{
The same IR with a \code{transform_type} column added to \code{projections}
and a new \code{stage_transforms} summary tibble.
}
\description{
Classify the transformations in a lineage IR
}
