% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sql.R
\name{sqlglot_available}
\alias{sqlglot_available}
\title{Is the Python \code{sqlglot} library available?}
\usage{
sqlglot_available()
}
\value{
\code{TRUE} if reticulate can see a \code{sqlglot} module, else \code{FALSE}.
}
\description{
Is the Python \code{sqlglot} library available?
}
